change evtnum "save" var to match var it saves
Change evtnum from 64 to 32 bits in sinsp::restart_capture_at_filepos() to match the variable it
is saving, sinsp::m_nevts.  (Not to be confused with the 64-bit m_nevts members found in some
other structures.)

+++ userspace/libsinsp/sinsp.cpp
@@ -905,7 +905,7 @@ void sinsp::restart_capture_at_filepos(uint64_t filepos)
 	//
 	// Backup a couple of settings
 	//
-	uint64_t evtnum = m_nevts;
+	uint32_t evtnum = m_nevts;
 	string filterstring = m_filterstring;
 
 	//
